requirements.txt
notion-client
pyperclip
keyboard
python-dotenv
